solution that doesnt break API
Change example and use a computed property so that each time called creates a new instance of the custom keyboard. To assure that a view always uses the same instance, use StateObject so that it stays the same obect throughout and the onSubmit from environment can be applied to the same custom keyboard instance. Adjust documentation accordingly